requirements.txt
fastapi
Pillow
numpy
python-multipart
uvicorn
